Changing the 'counting_sort.java' file to the Java folder
